pyproject.toml
[tool.poetry]
name = "justwork-trial"
version = "0.1.0"
description = ""
authors = ["Eduard Gan <eduard.gan@gmail.com>"]

[tool.poetry.dependencies]
python = "=3.8.7"
psycopg2-binary = "=2.8.6"
Django = "=2.2.18"
djangorestframework = "=3.12.2"
celery = "=5.0.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
